+++ package.json
@@ -4,15 +4,17 @@
     "version": "11.8.0",
     "description": "Lightweight JavaScript library for communicating with TAK Server",
     "author": "Nick Ingalls <nick@ingalls.ca>",
-    "main": "dist/index.js",
-    "types": "index.ts",
+    "main": "dist/index.node.js",
+    "types": "dist/index.d.ts",
     "bin": {
         "tak": "./cli.ts"
     },
     "exports": {
         ".": {
             "types": "./dist/index.d.ts",
-            "import": "./dist/index.js"
+            "react-native": "./dist/index.native.js",
+            "import": "./dist/index.node.js",
+            "default": "./dist/index.node.js"
         },
         "./lib/*": "./dist/lib/*.js",
         "./lib/api/*": "./dist/lib/api/*.js"
@@ -21,7 +23,7 @@
         "node": ">= 22"
     },
     "scripts": {
-        "test": "c8 --reporter=lcov ts-node-test test/",
+        "test": "vitest run",
         "lint": "eslint *.ts test/ lib/",
         "doc": "typedoc index.ts",
         "build": "tsc --build",
@@ -35,6 +37,7 @@
         "mime": "^4.0.7",
         "p12-pem": "^1.0.5",
         "pem": "^1.14.8",
+        "eventemitter3": "^5.0.1",
         "undici": "^7.8.0",
         "xml-js": "^1.6.11"
     },
@@ -56,17 +59,14 @@
         "@types/minimist": "^1.2.5",
         "@types/node": "^24.0.0",
         "@types/pem": "^1.14.4",
-        "@types/tape": "^5.6.0",
-        "c8": "^10.1.3",
         "eslint": "^9.0.0",
         "minimist": "^1.2.8",
-        "tape": "^5.6.1",
         "ts-node": "^10.9.1",
-        "ts-node-test": "^0.4.0",
         "tsx": "^4.19.4",
         "typedoc": "^0.28.1",
         "typescript": "^5.7.0",
-        "typescript-eslint": "^8.0.0"
+        "typescript-eslint": "^8.0.0",
+        "vitest": "^1.6.0"
     },
     "repository": {
         "type": "git",
