+++ test/findCoT.test.ts
@@ -1,39 +1,37 @@
+import { describe, it, expect } from 'vitest';
 import './setup.js';
 import TAK from '../index.js';
-import test from 'tape';
 
-test('findCoT - Unfinished', (t) => {
-    const res = TAK.findCoT('<event ><detail>');
-    t.equals(res, null);
-    t.end();
-});
-
-test('findCoT - Basic', (t) => {
-    const res = TAK.findCoT('<event></event>');
-    t.deepEquals(res, {
-        event: '<event></event>',
-        remainder: '',
+describe('findCoT', () => {
+    it('returns null when cot is unfinished', () => {
+        const res = TAK.findCoT('<event ><detail>');
+        expect(res).toBeNull();
     });
-    t.end();
-});
 
-test('findCoT - New Lines', (t) => {
-    const res = TAK.findCoT(`
+    it('parses basic event', () => {
+        const res = TAK.findCoT('<event></event>');
+        expect(res).toEqual({
+            event: '<event></event>',
+            remainder: '',
+        });
+    });
+
+    it('parses multiline event', () => {
+        const res = TAK.findCoT(`
 <event>
     <detail remarks="
 I am a multiline
 remarks field
     "/>
 </event>`);
-    t.deepEquals(res, {
-        event: '<event>\n    <detail remarks="\nI am a multiline\nremarks field\n    "/>\n</event>',
-        remainder: '',
+        expect(res).toEqual({
+            event: '<event>\n    <detail remarks="\nI am a multiline\nremarks field\n    "/>\n</event>',
+            remainder: '',
+        });
     });
-    t.end();
-});
 
-test('findCoT - New Lines - Non-Greedy', (t) => {
-    const res = TAK.findCoT(`
+    it('parses multiline non-greedy', () => {
+        const res = TAK.findCoT(`
 <event>
     <detail remarks="
 I am a multiline
@@ -45,48 +43,45 @@ I am a multiline
 remarks field
     "/>
 </event>`);
-    t.deepEquals(res, {
-        event: '<event>\n    <detail remarks="\nI am a multiline\nremarks field\n    "/>\n</event>',
-        remainder: '<event>\n    <detail remarks="\nI am a multiline\nremarks field\n    "/>\n</event>',
+        expect(res).toEqual({
+            event: '<event>\n    <detail remarks="\nI am a multiline\nremarks field\n    "/>\n</event>',
+            remainder: '<event>\n    <detail remarks="\nI am a multiline\nremarks field\n    "/>\n</event>',
+        });
     });
-    t.end();
-});
 
-test('findCoT - bad preceding data', (t) => {
-    const res = TAK.findCoT(`
+    it('handles leading garbage', () => {
+        const res = TAK.findCoT(`
 <fake/>
 <event><detail remarks="I am remarks"/>
 </event>
 `);
-    t.deepEquals(res, {
-        event: '<event><detail remarks="I am remarks"/>\n</event>',
-        remainder: '\n',
+        expect(res).toEqual({
+            event: '<event><detail remarks="I am remarks"/>\n</event>',
+            remainder: '\n',
+        });
     });
-    t.end();
-});
 
-test('findCoT - bad post data', (t) => {
-    const res = TAK.findCoT(`
+    it('handles trailing garbage', () => {
+        const res = TAK.findCoT(`
 <event><detail remarks="I am remarks"/>
 </event>
 <fake/>
 `);
-    t.deepEquals(res, {
-        event: '<event><detail remarks="I am remarks"/>\n</event>',
-        remainder: '\n<fake/>\n',
+        expect(res).toEqual({
+            event: '<event><detail remarks="I am remarks"/>\n</event>',
+            remainder: '\n<fake/>\n',
+        });
     });
-    t.end();
-});
 
-test('findCoT - mixed', (t) => {
-    const res = TAK.findCoT(`
+    it('handles mixed data', () => {
+        const res = TAK.findCoT(`
 <event><detail remarks="I am remarks"/>
 </event>
 <fake/>
 <event><detail remarks="I am remarks"/></event>`);
-    t.deepEquals(res, {
-        event: '<event><detail remarks="I am remarks"/>\n</event>',
-        remainder: '\n<fake/>\n<event><detail remarks="I am remarks"/></event>',
+        expect(res).toEqual({
+            event: '<event><detail remarks="I am remarks"/>\n</event>',
+            remainder: '\n<fake/>\n<event><detail remarks="I am remarks"/></event>',
+        });
     });
-    t.end();
 });
